Read more boot data
Since the second console used on s390x is sclp/sclplm, then it will get
also the boot messages: hence, read 8K of data each time, to be able to
drain them all, and get the announce payload within a short waiting
time.  This change the behaviour on the other architectures as well, but
it should be harmless.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>