Optionally use libosinfo for OSes URLs
Try to use libosinfo (if available) to get the URLs of install trees
and ISOs of the supported OSes, using a libosinfo short ID provided by
all the OS classes.  Add <osinfo> and <variant> child tags of <os> to
override (or set) the short ID, and its libosinfo variant.

This makes the <url> and <iso> child tags of <install> optional, as they
can be determined automatically from libosinfo, at least from the
majority of cases.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>