Add location for newer edk2 OVMF binaries
The newer upstream of edk2 (Tianocore) ships the OVMF binaries in a new locaiton
with the older /usr/share/OVMF binaries often symlinked to the new path. Add
the newer defaults so we'll still work if the symlinks ever go.

Add 32 bit ARM ARMv7 as an uEFI architecture

For some time now 32 bit ARMv7 has been supported in edk2 so enable support
for detecting the binaries if they happen to be available to support using
uEFI on ARMv7 and related.

Add a USB controller and keyboard on ARMv7/aarch64 VMs

Due to some historical reason the ARM hosts never added a USB controller or a
keyboard and as a result the grpahical console doesn't work. So add them as
to the VM so the graphical console works and we can capture screen shots if
an install fails to aid in debugging like other architectures.

Check for uEFI on ARMv7/x86_64 as well as just aarch64

Enable checking for uEFI edk2 firmware on all architectures that support it,
this enables OS installers to detect uEFI and build images supporting it and
hence be able to support things like secure boot on devices that enable it.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>