Guest: fix to always use host-passthrough CPU for all arches
The use of host passthrough was restricted to just arm architectures,
which meant other arches got their built in default, which is almost
never what you want. eg x86 gets qemu64 which lacks so many features
that some guests will not function at all.

There is no need to care about live migration with Oz, so it should
unconditionally use host-passthrough for all architectures when KVM
is enabled.

Furthermore the way host-passthrough was requested was incorrect
and relying on an accident of the libvirt QEMU impl. It needs to
use mode=host-passthrough, not mode=custom + model=host. Libvirt
should block the latter from being used.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>