Unpick unnecessary 'useuefi' arg on the Guest classes
b7728d5 added this new mandatory argument on the Guest class. But
it doesn't use it at all. The only code in the commit that uses
anything called 'useuefi' is the little bit added to
_generate_xml() - but that checks self.config.useuefi, which is
a property of the FedoraConfiguration config class, not anything
to do with the Guest class itself. None of the __init__ methods
on Guest or its subclasses that added this new 'useuefi' arg ever
did anything with it.

This broke every guest class which didn't have the useless new
argument added to it, which seems to be most of the non-Fedora/
non-RHEL ones. They're probably broken in other ways anyway,
but this didn't help. It makes a large swathe of the test suite
fail.

This dumps the useless arg from all Guest classes, and makes the
check in _generate_xml() safe for config classes which do not
have the useuefi property at all, which is all the non-Fedora
ones.

Signed-off-by: Adam Williamson <awilliam@redhat.com>