Add ability to run the unit tests in Fedora and EL7 containers
This adds a convenient way to run the unit tests in containers -
a "latest Fedora" container to check things work with the latest
shiny bits, and an EL 7 (CentOS 7 plus EPEL) container to check
things work with antique bits. This should make it possible to
conveniently set up CI with a github action. It also should
make it easy to run the tests on non-Fedora/RHEL-ish platforms.

`make container-unittests` will build both containers and run
the tests. `make container-clean` cleans up the containers and
images.

Signed-off-by: Adam Williamson <awilliam@redhat.com>