CI: use sudo, assume docker present, use diff-{quality,cover}
Huh. I kinda just assumed we'd be root. Guess not!

Docker is already in the GHA base image, no need to install.

Both linter tools report issues on the current master branch,
which means they will always fail on exit. Using diff-quality
gives us a check we can usefully fail on (as configured in this
PR, it will only fail if the quality of the lines changed in
the PR is less than 9.0).

Similarly we can generate a coverage report and use diff-cover
to fail if coverage of the PR is under 90%.

This also uses a standard configuration file name for the pylint
config (so we don't have to specify it with `--rcfile`), removes
all non-default settings from the pylint config so it's clearer
what we really intend to configure, and moves the flake8 config
to a config file like pylint.

We don't use pyproject.toml for configuration here because GHA's
Ubuntu environment is just not new enough, the tools in it don't
consistently read from pyproject.toml.

Signed-off-by: Adam Williamson <awilliam@redhat.com>